+++ docs/en/docs/release-notes.md
@@ -5,6 +5,16 @@ hide:
 
 # Release Notes
 
+## 0.2.4
+
+### Changed
+
+- [Query Parameters](./extras/query-params.md) documentation updated with generics (list, dict) examples.
+
+### Fixed
+
+- Parsing query parameters with multiple datastructures was not parsing it properly
+
 ## 0.2.3
 
 ### Added